requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
pg8000==1.30.2
python-dotenv==1.2.2
Flask
flask_sqlalchemy
psycopg2-binary
python-dotenv
gunicorn